refactor(): removed a few obsolete files from src root, moved other things around
