+++ src/old/parseStlAsStreamWorker.js
