+++ src/parseStreamAlt.js
@@ -2,18 +2,27 @@ import { isDataBinaryRobust, ensureBinary, ensureString } from './utils'
 
 export default function makeStlStreamParser () {
   let chunkNb = 0
-  let totalSize = 0
   let isBinary = false
 
   let faceOffset = 0
   let previousRemainderData
 
   const parser = function (chunk, enc, callback) {
-    //console.log('chunk', chunk.length, chunkNb)
+    // console.log('chunk', chunk.length, chunkNb)
     if (chunkNb === 0) {
-      isBinary = isDataBinaryRobust(chunk.buffer)
+      isBinary = isDataBinaryRobust(chunk)
     }
-    const workChunk = ensureBinary(previousRemainderData ? Buffer.concat([previousRemainderData, chunk], previousRemainderData.length + chunk.length) : chunk)
+
+    let workChunk
+    if (previousRemainderData) {
+      workChunk = new Uint8Array(previousRemainderData.byteLength + chunk.byteLength)
+      workChunk.set(new Uint8Array(previousRemainderData))
+      workChunk.set(new Uint8Array(chunk), previousRemainderData.byteLength)
+      workChunk = workChunk.buffer
+      //let data = ensureString(workChunk)
+    } else {
+      workChunk = chunk
+    }
 
     const {remainingDataInChunk, startOffset} = isBinary ? computeBinaryOffset(workChunk, chunkNb) : computASCIIOffset(workChunk, chunkNb)
 
@@ -22,35 +31,19 @@ export default function makeStlStreamParser () {
 
     // update size
     chunkNb += 1
-    totalSize += chunk.length
 
     previousRemainderData = parsed.remainderData
     faceOffset = parsed.faceOffset
 
-    // we can only send a single buffer out , so concat the two
-    const positionsBuffer = toBuffer(parsed.positions)
-    const normalsBuffer = toBuffer(parsed.normals)
-    // console.log('positions', positionsBuffer.length, parsed.positions.length)
-    //console.log('done with chunk', positionsBuffer, callback)
-    callback(null, Buffer.concat([positionsBuffer, normalsBuffer], positionsBuffer.length + normalsBuffer.length))
+    callback(null, parsed.data)
   }
 
   return parser
 }
 
-// helper function, taken from https://github.com/feross/typedarray-to-buffer
-function toBuffer (arr) {
-  var buf = new Buffer(arr.buffer)
-  if (arr.byteLength !== arr.buffer.byteLength) {
-    // Respect the "view", i.e. byteOffset and byteLength, without doing a copy
-    buf = buf.slice(arr.byteOffset, arr.byteOffset + arr.byteLength)
-  }
-  return buf
-}
-
 function computeBinaryOffset (workChunk, chunkNb) {
   const dataStartOffset = 84
-  const remainingDataInChunk = chunkNb === 0 ? workChunk.length - dataStartOffset : workChunk.length
+  const remainingDataInChunk = chunkNb === 0 ? workChunk.byteLength - dataStartOffset : workChunk.byteLength
   const startOffset = chunkNb === 0 ? dataStartOffset : 0
 
   return {remainingDataInChunk, startOffset}
@@ -66,12 +59,13 @@ function parseBinaryChunk (faceOffset, remainingDataInChunk, startOffset, workCh
   const faceLength = 12 * 4 + 2
   const facesInChunk = Math.floor(remainingDataInChunk / faceLength)
 
-  const reader = new DataView(workChunk.buffer)
+  const reader = new DataView(workChunk)
   const faces = reader.getUint32(80, true)
-  let positions = new Float32Array(facesInChunk * 3 * 3)
-  let normals = new Float32Array(facesInChunk * 3 * 3)
+  // let positions = new Float32Array(facesInChunk * 3 * 3)
+  // let normals = new Float32Array(facesInChunk * 3 * 3)
 
-  let data = new Float32Array(facesInChunk * 2 * 3 * 3)//we want one big typedArray for all the data , to avoid concat operations
+  const halfLength = facesInChunk * 3 * 3
+  let data = new Float32Array(facesInChunk * 2 * 3 * 3) // we want one big typedArray for all the data , to avoid concat operations
 
   let lface = 0
   let offset = 0
@@ -82,22 +76,39 @@ function parseBinaryChunk (faceOffset, remainingDataInChunk, startOffset, workCh
     let start = startOffset + lface * faceLength
     for (var i = 1; i <= 3; i++) {
       let vertexstart = start + i * 12
-      positions[ offset ] = reader.getFloat32(vertexstart, true)
-      positions[ offset + 1 ] = reader.getFloat32(vertexstart + 4, true)
-      positions[ offset + 2 ] = reader.getFloat32(vertexstart + 8, true)
+      data[ offset ] = reader.getFloat32(vertexstart, true)
+      data[ offset + 1 ] = reader.getFloat32(vertexstart + 4, true)
+      data[ offset + 2 ] = reader.getFloat32(vertexstart + 8, true)
 
-      normals[ offset ] = reader.getFloat32(start, true)
-      normals[ offset + 1 ] = reader.getFloat32(start + 4, true)
-      normals[ offset + 2 ] = reader.getFloat32(start + 8, true)
+      data[ offset + halfLength ] = reader.getFloat32(start, true)
+      data[ offset + halfLength + 1 ] = reader.getFloat32(start + 4, true)
+      data[ offset + halfLength + 2 ] = reader.getFloat32(start + 8, true)
       offset += 3
     }
     lface += 1
   }
 
   // compute offsets, remainderData etc for next chunk etc
-  const remainderData = workChunk.slice(workChunk.length - remainingDataInChunk % faceLength) // chunk length - remainderDataLength
+  const remainderData = workChunk.slice(workChunk.byteLength - remainingDataInChunk % faceLength)
+  // workChunk.slice(workChunk.length - remainingDataInChunk % faceLength) // chunk length - remainderDataLength
 
-  return {faceOffset: lface, remainderData, positions, normals}
+  return {faceOffset: lface, remainderData, data}
+}
+
+// taken from http://stackoverflow.com/questions/6965107/converting-between-strings-and-arraybuffers
+function str2ab (str) {
+  var buf = new ArrayBuffer(str.length * 2) // 2 bytes for each char
+  var bufView = new Uint16Array(buf)
+  for (var i = 0, strLen = str.length; i < strLen; i++) {
+    bufView[i] = str.charCodeAt(i)
+  }
+  return new Uint8Array(buf)
+  /*var idx, len = str.length, arr = new Array( len );
+   for ( idx = 0 ; idx < len ; ++idx ) {
+       arr[ idx ] = str.charCodeAt(idx) & 0xFF;
+   }
+   // You may create an ArrayBuffer from a standard array (of values) as follows:
+   return new Uint8Array( arr )*/
 }
 
 // ASCII stl parsing
@@ -117,7 +128,6 @@ export function parseASCIIChunk (workChunk) {
 
   while ((result = patternFace.exec(data)) !== null) {
     var length = 0
-
     text = result[0]
     offset = result.index + text.length
 
@@ -134,15 +144,19 @@ export function parseASCIIChunk (workChunk) {
     faces += 1
   }
 
-  let positions = new Float32Array(faces * 3 * 3)
+  /*let positions = new Float32Array(faces * 3 * 3)
   let normals = new Float32Array(faces * 3 * 3)
 
   positions.set(posArray)
-  normals.set(normArray)
+  normals.set(normArray)*/
+  let outData = new Float32Array(faces * 3 * 3 * 2)
+  outData.set(posArray)
+  outData.set(normArray, posArray.byteLength)
 
   // compute offsets, remainderData etc for next chunk etc
   const remainderTextData = data.slice(offset)
-  // console.log('remainderData', remainderTextData)
-  const remainderData = new Buffer(remainderTextData)
-  return {faceOffset: faces, remainderData, positions, normals}
+  const remainderData = str2ab(remainderTextData)
+  let check = ensureString(remainderData)
+  //console.log('remainderData', remainderTextData, remainderData)
+  return {faceOffset: faces, remainderData, data:outData}
 }
